pyproject.toml
[project]
name = "cloudflare-personalization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.34.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
